Updated Files To Change Player SlotID
These updateWeek.php page has been updated to use GET rather than POST
so that the updateWeekScript.php can make it run through cURL.
updateWeekScript can be used in conjunction with a CRON job to work on a
weekly basis. It will in this order, move over the fantasy points of
each player scored that week (those players that did not play are
ignored), update the SlotID in for each player in the weekplayerdata
table, and then finally clear all statistics for the entire
weekplayerdata table.