New Folder To Hold SQL Commands For DB Cleanup
The two files I've committed into this new folder clean up the database.
The inapplicable positions.sql file removes all players that are not QB,
K, WR, Flex, or RB positions. The ResetAuctionValues.sql resets the
weekplayerdata and seasonplayerdata current auction values to the
initial. It also resets the fantasy points and stats to 0 in the
weekplayerdata