Insure Copies Of Everything Made It To Github
This is my entire capstone directory from my personal babbage space that
I've been working for all the simulation, database interaction, json
file creation, etc for the entire semester. Copies exist within the rest
of the repository, but I wanted to insure nothing was missed, and there
is a final copy of everything available to everyone. This folder DOES
NOT affect the rest of the project and should be only used to look for
resources that may not have been filed appropriately